Work around long update check interval in controller with a restart
Previously, the base/proxy-and-update e2e test passed, because
the controller considered `latest < {booted,inactive}` an error state
and triggered an update check "quickly" (in 30 seconds).

This was changed with commit d8053bd0a7da78e8e20dd43440fbcf8d87e71ae9
which changed the interpration of this state to `UpToDate` i.e. a
non-error state, therefore entering a long sleep of 1 hour.

As a work-around, we restart the controller to trigger an update check.
This is ugly and potentially hides bugs, should be eliminated once
the runtime configuration option is added.