Split Update state into version info, system status and process state
This refactors Update state into 3 parts:
- `process_state`, which indicates what action the Update process is
  currently performing (checking, downloading, installing or sleeping)
- `system_status`, which reflects what the Update process thinks is the
  current system state (up-to-date, needs-update, etc)
- `version_info`, which reflects the current slot and latest upstream
  versions.

The reason for doing this is to decompose the process transitions from
the (determined) system state, as consumed by other components via the
`React.Signal`. For example, this allows to observe the last
`system_status` in the GUI while the Update process is performing other
tasks.

Note that only the `process_state` is an input state (meaning that the
behaviour of Update process depends on it), whereas the
`system_status` and `version_info` are "output" states.

In theory, the basic FSM step could be refactored into something like:

        val run_step: process_state
            -> (process_state * (system_status option) * (version_info option)) Lwt.t

where the recursive loop would now `run_rec` maintain the `state` and
update it's fields if the output `option`s are set. However, this splits the
logic over two functions rather than one and makes it more complicated
to interpret state machine.