Remove HDMI from module-switch-on-connect blacklist
nixpkgs 24.11 bumps pulseaudio version from 16.1 to 17.0 which brings
a major change in ALSA -> PulseAudio device mapping:

    An extensive set of changes landed which modify how ALSA UCM
    configuration is mapped to PulseAudio profiles and ports. Notably:
    <..> Instead of different inputs/outputs being exposed as ports on a
    source/sink, they *will be exposed as separate sources/sinks*.

From the changelog:
https://www.freedesktop.org/wiki/Software/PulseAudio/Notes/17.0/#updatestoalsaucm-basedsetups)

This seems to imply that a HDMI cable plug/unplug, which was previously
a port change and hence handled by module-switch-on-port-available,
is now a sink change (and hence handled by module-switch-on-connect).

However, since module-switch-on-connect by default blacklists any
devices with "hdmi" in their name, this would not trigger a change.