Make release disk a separate build target
Earlier release disk build had a critical problem: the built disk would
have optimized partition sizes based on the current system's image size.
This meant that newer releases would potentially not fit in the system
partition if they increase in size.

To avoid this issue, we create disk images which have system partition
sizes equal to (default PlayOS size - 1 GiB). The 1GiB is subtracted as a
safety threshold to avoid releasing something that is very near max
capacitiy.

This gets rid of the extra top-level `extraModules` argument in
default.nix along the way.