Fix hostapd usage in tests and ./build vm
Something changed with hostapd/network initialization after the latest
nixpkgs bump and hostapd was failing to create the simulated APs because
the wlan0 radio was soft-blocked with rfkill upon startup.

Additionally, the service ordering with connman seems to cause weird
issues, not sure if this always existed or is new.