Disable USB soundcards
We want to be able to assume a fixed set of soundcards for the standard
devices, and don't currently need any extensibility for additional
inputs or outputs.

Disable all support for USB soundcards. This fixes an issue where
USB peripherals could lead to unexpected numbering of sound cards and
prevent successful configuration of sound output to HDMI.