Enable mouse cursor hiding when idle
This lets the mouse cursor be initially hidden, appear when moved, and
hidden again after 10 seconds of being idle.

The timing is chosen so that the cursor is normally hidden, but stays
visible for relatively long period should someone attach and use a
mouse.

We have been hiding idle cursors in the Play web app, but have found
that the mouse needs to be moved once upon boot for the web app's logic
to work, which leaves the cursor permanently visible in mouse free use.
This seems to be a system/engine level issue, that we can not control
from the web app. By moving mouse hiding to the OS we can also remove
the complexity and event handling overhead from the web runtime.