Fix "input method is not set" errors after focusing field with new type
Focusing from a numeric to a text field seems to somehow "unset" the
internal inputMethod in Qt virtual keyboard, which prevents further
input.

Note: this is not due to the inputMethodHint manipulation in
inputpanel.qml, tried removing it and the error still happens.

Smells like a Qt bug.  The error message originates from here:
https://github.com/qt/qtvirtualkeyboard/blob/f909c0d09523106787418941d951c56b65680b26/src/virtualkeyboard/qvirtualkeyboardinputengine.cpp#L59