Fix Escape key handling by using regular keyReleaseEvent
With QShortcut there's weird interactions with the webview and no way to
stop event propagation. Partially because "Escape" is not really a
shortcut, it's a single key.

The Escape is now handled BrowserWidget, since it is the deepest child
widget and sees the event first, so it can prevent it from bubbling up
to the DialogableWidget.