Use an alternative work-around for "input method is not set" errors
The previous workaround used a crude approach of enabling
QT_VIRTUALKEYBOARD_FORCE_EVENTS_WITHOUT_FOCUS, but after further
inspection, this can lead to hard-to-debug issues. For example,
it causes the virtual keyboard to choose the totally arbitrary first
window as the target of keypress events, see here:
https://github.com/qt/qtvirtualkeyboard/blob/f909c0d09523106787418941d951c56b65680b26/src/virtualkeyboard/platforminputcontext.cpp#L213-L215

Instead, we resolve the issue "closer" to where it originates by
ensuring the input method is always populated. Since we only use
the PlainInputMethod this workaround is sufficient.

For future generations that might attempt to debug this, enable virtual
keyboard debug logs with

        export QT_LOGGING_RULES="qt.virtualkeyboard.debug=true"

and observe that QVirtualKeyboardInputEngine::setInputMethod is always
called with a non-null object, yet somehow the reference ends up
pointing to null. This is very confusing and seems to imply that
the input method instance gets destroyed (and hence the QPointer
pointing to is cleared). Yay, C++!