Use show/hide instead of resizing to 0 for vkb display
While attempting to reproduce the "input method is not set" for a Qt bug
report, I discoverd it is (in part) caused by resizing the
KeyboardWidget to 0 pixels (`resize(QSize(0, 0))`).

Using regular show()/hide() methods fixes the issue and is also a
more clean approach, since it does not require the input metod reset and
there is less needless resizing.

In addition, it seems to remove the visual jump/flicker when the
virtual keyboard is being repositioned bottom->top.