Patch connman to make service sorting demote link-local
This applies a patch to connman to make its service sorting factor in
the type of IPv4 method used by services. Link-local/Auto-IP configured
services are demoted relative to connected services using other methods.

This happens relatively late in the comparison hierarchy so that

- online services still win over ready services,
- ready services still win over failed-online-check services,
- preferred technologies still win over less preferred (which may be
  questionable, but also in line with current treatment of preferred
  technologies).

As link-local configured services are in READY state and remain in
READY state when receiving an address via DHCP, updating the comparison
function alone is not sufficient, we also need to add a new trigger for
re-sorting the service list when the IP address changes.

The chosen trigger here is address_updated(), which is called both when
the ipconfig method is changed manually as well as when the ipconfig
changes bottom up when DHCP becomes active again after the IP device had
fallen back to link-local/AUTO.

Considerations for this choice of re-sorting trigger:

- Suitability: Triggering on address updates seems appropriate given
  that we are ultimately interested in connectedness, not the ipconfig
  method alone. We want to prefer READY DHCP services in particular, not
  just any service for which DHCP is currently used but has not yieled
  an IP.
- Frequency: relatively low, compared with trigger from
  __connman_service_update_from_network() which triggers on average
  every 3 seconds on my laptop[^1]

[^1]: journalctl -u connman --since "-$(( 15 * 60 ))" | grep service_list_sort | grep -c service_update_from_network