Disable obsoleted ConnMan features
Both PreferredTechnologies and EnableOnlineCheck were used with the aim
of avoiding accidental default service use of network devices connected
only to a Senso.

PreferredTechnologies triggers a bug in ConnMan up to version 1.45[^1]
and yields to a possibly surprising online check bypass even with the
bug fixed.

The online check used in ConnMan up to version 1.42 (which we currently
pin to) is one-shot, meaning it will only help to hold on to an online
service once chosen as default service, but will not help in choosing
an online-capable service as default in the first place.

Both should be obsolete with ConnMan's service order updated with a
heuristic to demote link-local services relative to READY/ONLINE
services with other ipconfig methods which are more promising to yield a
functioning default route (1998317d861c1a967638ecfa4a823e852ecc32bf).

To still ensure that WiFi and Ethernet are always used in parallel when
available, PreferredTechnologies must be "replaced" to configure
parallel use without influencing the service ordering:

- `AlwaysConnectedTechnologies` to ensure both WiFi and Ethernet are
  always connected to if available (instead of one being considered
  "enough", I suppose)[^2]

In the following the defaults would have been aligned, but included
to pin and be explicit here:

- `DefaultAutoConnectTechnologies` to ensure that _known_ WiFis will
  automatically be connected to, without requiring user action, for
  example to automatically connect upon reboot[^3]
- `DefaultFavoriteTechnologies` being set to `ethernet` means that
  for Ethernet, even _unkown_ networks will be connected to
  automatically.

[^1]: https://git.kernel.org/pub/scm/network/connman/connman.git/commit/src/service.c?id=82ce27e45308cf2478320bdb2a63c23efda35d0d
[^2]: https://git.kernel.org/pub/scm/network/connman/connman.git/tree/doc/connman.conf.5.in?h=1.42#n86
[^3]: https://git.kernel.org/pub/scm/network/connman/connman.git/tree/doc/connman.conf.5.in?h=1.42#n73