Add an option to label networks with services of interest
This optionally lets the controller browse for services of one or more
types via DNS-SD and annotates connected networks with the instance name
of discovered services.

This is introduced in order to be able to mark the network service that
is connected to a Senso in the list, instead of just having two "Wired"
services and relying on IP to differentiate which one is more likely to
"have" a Senso.

So for example, the Senso announces itself with service type
`_sensoControl._tcp` and instance name `Dividat Senso`. When the
controller activates annotations for this service type, a network on
which a Senso advertises itself is going to be labelled with "Dividat
Senso" in the network list.