Browse specific service types for network labels
Avahi's cache is specific to queried service types, and under the hood,
running `avahi-browse --all` queries a special service type for service
enumeration (`_services._dns-sd._udp.local`)[^1]. This means that
queries for specific service types and `--all` queries will not work
against the same cache.

So in order to dump the same cache in the OCaml bindings that we prefill
with the services, we need to query by service type in both cases.

The alternative would be to align both places to use `--all` queries,
but this is not actually necessary and enumerating services on all
interfaces might concern attentive network administrators.

[^1]: https://datatracker.ietf.org/doc/html/rfc6763#section-9