Toggle virtual keyboard visibility with explicit user intercation
With these changes, the keyboard is no longer automatically
displayed/hidden on input field focus/unfocus, but requires the user
to perform an action (Enter/OK to open, Esc/Hide button to close it).

All key handling is refactored to happen via the top-level QApplication
instance to avoid other Qt objects from intercepting the key events.

The implementation painstakingly attempts to distinguish explicit
keyboard hide/close actions from "implicit" platform hide/close actions
(due to navigation or otherwise), as a means of tracking the state of
(potential) input. This could be vastly simplified if Qt provided
a signal like inputObjectChanged to be able to know if the user is
currently focused on an input field or not.