Track focus object to avoid using event filters on QApplication.instance()
Installing an eventFilter on QApplication can degrade the performance,
so it is best avoided, even if more convenient.

Basic profiling shows that such eventFilter's are called ~500 times/sec and
consume ~20ms of CPU time per second (2%), even if mostly not doing anything.
This is a bit too much to be called negligible.

This slightly changes what events are visible, since it seems
ShortcutOverride events do not arrive for focusObject if there are no
(Qt) Shortcuts using the the keys defined for it.