Hygiene: avoid using Qt signal handlers not bound to objects
Signal handlers (i.e. Slots in Qt's terms) not bound to an object
instance do not get automatically disconnected and GC'ed, which can
lead to crashes where the handler outlives the object it should belong
to.

The cases here should not lead to crashes within the current usage, but
it's good to avoid these patterns in case the code is refactored.