Make password reveal button reachable via spatial navigation
The password SHOW/HIDE button in network configuration interfaces was
not reachable with spatial navigation, because it was located entirely
within the bounding box of the input element it controls visibility off.

The simplest way to enable spatial navigation is just to locate the
exact same button element to the right of the input field. The padding
on the input is then not needed.