Use QWebChannel to set focus-shift interaction mode on keyboard change
This script, which is injected into all pages including the main kiosk
page, sets a CSS variable on the document element (`<html>`) to control
whether focus-shift treats inputs as opaque or normal, i.e. allows arrow
key interaction.

The QWebChannel allows the front-end code to observe/subscribe to the same
keyboard_available_changed events as the backend.

When there is a physical keyboard present we use the normal interaction
behavior, when there is no physical keyboard we disable special behavior
on inputs.

Co-authored-by: Ignas Vyšniauskas <ignas@dividat.ch>