Use focus-shift version with opaque option
This uses an untagged main-branch version of focus-shift[^1], with
support for configuring opaque behavior on inputs, which the kiosk sets
via a QWebChannel when only a remote control is found.

We may bump to a tagged release of focus-shift after a few more feature
additions in the lib have been completed.

[^1]: https://github.com/dividat/focus-shift/commit/1e43419e1562dbe10ab4ce2a768e0fd0d51148bb