Make reveal toggle button explicitly unrelated to form
With a plain button, which is implicitly `type="submit"` its onclick
handler is called when the form is submitted. Leading in this case to
the form not being submitted and password visibility toggled instead
when pressing 'Enter' in the password input field.

The fix is to annotate the button as just `type="button"`, which means
it no longer has a default action we need to prevent either.