Let password reveal toggle always be visible
In my experience it is both comprehensible as well as a better flow to
be able to reveal password before starting to enter it.

With this change, it also no longer makes sense to auto-hide the
password when the field is edited from being empty, as the user may just
have activated the button to reveal the password.