Apply a ConnMan patch for less aggressive bgscan
I am not sure why this is hardcoded upstream, but I think ConnMan's
defaults are really not appropriate for stationary clients with a need
for reliable connectivity over micro-optimizing WiFi speed.

The aggressive scanning and AP switching strategy makes random
connection errors more likely, which is problematic when paired with
ConnMan's tendency to give up on services with errors rather quickly.