Add VM-based test for SAE retry mechanism
This is a very basic sanity check to confirm that the retry mechanism is
indeed triggered, though we can not test successful retries with this
simple test. Full testing of the intermittent failure with eventual
success scenario would require controlling hostapd to simulate the
required events for ConnMan to process.