Wrap kiosk in a cgroup with a MemoryHigh limit
Recent experiments have shown that chromium (and hence QtWebEngine) will
keep consuming close to max of the available memory, making it more
likely that the OOM killer will get involved due to a spike in memory
usage. The spike can be caused both by chromium itself (e.g. loading a
heavy page) or by routine system activities (e.g. downloading of new
PlayOS bundle).

To avoid this, we start kiosk in a cgroup (via a transient systemd
scope) with constrained resources. Setting MemoryHigh < MemoryMax
ensures that chromium will initiate GC before the system reaches a
critical state.

Unfortunately, unlike for services, systemd does not provide an
`ExitType=cgroup` behaviour for scopes. Meaning it is not possible to
declaratively specify that all processes in the cgroup need to be
terminated when the scope exits. As a workaround, we perform manual
cleanup on exit.