Handle play:beforereload as a "full" reload of in kiosk
The goal of this action is to give a hint to QWebEngineView/chromium to
drop some cached items from memory and perform "deeper" garbage
collection.

JS heap GC, which could be exposed to the app via `--expose-gc`, is not
sufficient for our needs and as of now there are no alternative APIs to
induce GC.

This is based on empirical observations that switching between Play and
PlayOS settings views results in a big memory reduction in kiosk, so we
simulate that, in addition to a "conventional" reload.