Set custom os-release meta data
Sets PlayOS specific meta data in /etc/os-release, which systemd
variously relies on.

Considered but not added for now:

- VENDOR_NAME, VENDOR_URL
- SUPPORT_URL, DOCUMENTATION_URL
- ANSI_COLOR -- no color, no bikeshedding :-)

We don't rely on these in our UI, and to my understanding nothing is
displayed in terminal outputs if these are not set.

Setting custom meta data in this file also disables the EOS warning
during boot we previously inherited from the underlying NixOS version.