Delete raucb files after any install outcome
We previously deleted downloaded raucb files after a failed install, but
not if installation succeeded. This may have been done assuming that the
system is eventually rebooted post-install. While this clears the
download in the long run, this may take days or weeks after
installation, which is especially problematic because the download
happens to tmpfs.

With this change, the bundle removal is performed no matter the
installation outcome.