Avoid /boot/status.ini corruption in release-validation for legacy systems
QEMU's system_reset does an "unclean" reboot (power plug), which causes
corruption of the FAT filesystem in the test. There is no way to fix
this for existing systems and their DISK images, so instead we try to
work-around the issue by giving time for an unmount/fsync to happen.