Check that system status is marked Good after update
A "naive" check fails currently, because /boot/status.ini gets
corrupted due an (unclean) system_reset "soon" after an install (which
is the equivalent of a power plug).  The existing statusfile-recovery
does not seem to be sufficient. It might be due to this being run in a
VM and/or due to using a virtual file-backed disk, but illustrates a
problem with using a FAT filesystem for persisting RAUC's status file.

Unfortunately, there is no way to trigger a "clean" reboot and no easy
way to fix this for existing systems and their DISK images, so instead
we work-around the issue by giving time for an unmount/fsync to happen.