Let systemd activate the hardware watchdog
This lets systemd activate the hardware's watchdog device, meaning
systemd will control and periodically touch the watchdog. If systemd or
even the Linux kernel should end up crashed or frozen, the watchdog
would perform a reset at the hardware level, such that the system
reboots. systemd pings the watchdog at half the set interval.

We use the default watchdog device, `/dev/watchdog0`.

This is the minimum configuration to utilize the watchdog with systemd.
Beyond this, if there is any concern that individual services may fail
to function while the system is generally operational (but maybe in a
bad state), systemd also allows integrating individual services with the
watchdog, such that both repeated failure of the service and failure to
call the watchdog will trigger a system reboot.

https://www.freedesktop.org/software/systemd/man/latest/systemd-system.conf.html#Hardware%20Watchdog
https://www.freedesktop.org/software/systemd/man/latest/systemd.service.html#WatchdogSec=
https://0pointer.de/blog/projects/watchdog.html